Replace class with interface
